Crate fixed_decimal

fixed_decimal is a utility crate of the ICU4X project.

This crate provides Decimal and UnsignedDecimal, essential APIs for representing numbers in a human-readable format. These types are particularly useful for formatting and plural rule selection, and are optimized for operations on individual digits.

Examples

use fixed_decimal::Decimal;

let mut dec = Decimal::from(250);
dec.multiply_pow10(-2);
assert_eq!("2.50", format!("{}", dec));

#[derive(Debug, PartialEq)]
struct MagnitudeAndDigit(i16, u8);

let digits: Vec<MagnitudeAndDigit> = dec
    .magnitude_range()
    .map(|m| MagnitudeAndDigit(m, dec.digit_at(m)))
    .collect();

assert_eq!(
    vec![
        MagnitudeAndDigit(-2, 0),
        MagnitudeAndDigit(-1, 5),
        MagnitudeAndDigit(0, 2)
    ],
    digits
);

Structs

CompactDecimal

A struct containing a Decimal significand together with an exponent, representing a number written in compact notation (such as 1.2M). This represents a source number, as defined in UTS #35. The value exponent=0 represents a number in non-compact notation (such as 1 200 000).

FixedInteger

A FixedInteger is a Decimal with no fractional part.

LimitError

The magnitude or number of digits exceeds the limit of the UnsignedDecimal or Decimal.

ScientificDecimal

A struct containing a Decimal significand together with an exponent, representing a number written in scientific notation, such as 1.729×10³.

Signed

The Signed struct represents a numeric value with an associated sign.

UnsignedDecimal

A struct containing decimal digits with efficient iteration and manipulation by magnitude (power of 10).

Enums

ParseError

An error involving FixedDecimal operations or conversion.

RoundingIncrement

Increment used in a rounding operation.

Sign

A specification of the sign used when formatting a number.

SignDisplay

Configuration for when to render the minus sign or plus sign.

SignedRoundingMode

Mode used in a signed rounding operations.

UnsignedRoundingMode

Mode used in a unsigned rounding operations.

Type Aliases

Decimal

A Type containing a UnsignedDecimal and a Sign to represent a signed decimal number.