Struct fixed_decimal::decimal::FixedDecimal

pub struct FixedDecimal { /* private fields */ }

A struct containing decimal digits with efficient iteration and manipulation by magnitude (power of 10). Supports a mantissa of non-zero digits and a number of leading and trailing zeros, as well as an optional sign; used for formatting and plural selection.

Data Types

The following types can be converted to a FixedDecimal:

Integers, signed and unsigned

Strings representing an arbitrary-precision decimal

To create a FixedDecimal with fraction digits, either create it from an integer and then call FixedDecimal::multiplied_pow10, or create it from a string.

Floating point numbers will be supported pending a resolution to #166. In the mean time, a third-party float-to-string library may be used.

Magnitude and Position

Each digit in a FixedDecimal is indexed by a magnitude, or the digit’s power of 10. Illustration for the number “12.34”:

Magnitude	Digit	Description

1	1	Tens place

0	2	Ones place

-1	3	Tenths place

-2	4	Hundredths place

Some functions deal with a position for the purpose of padding, truncating, or rounding a number. In these cases, the position is the index on the right side of the digit of the corresponding magnitude. Illustration:

Position:   2   0  -2
Number:     |1|2.3|4|
Position:     1  -1

Expected output of various operations, all with input “12.34”:

Operation	Position	Expected Result

Truncate Left	1	10

Truncate Right	-1	12.3

Pad Left	4	0012.34

Pad Right	-4	12.3400

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from(250);
assert_eq!("250", dec.to_string());

dec.multiply_pow10(-2);
assert_eq!("2.50", dec.to_string());

Implementations

impl FixedDecimal

pub fn digit_at(&self, magnitude: i16) -> u8

Gets the digit at the specified order of magnitude. Returns 0 if the magnitude is out of range of the currently visible digits.

Examples

use fixed_decimal::FixedDecimal;

let dec = FixedDecimal::from(945);
assert_eq!(0, dec.digit_at(-1));
assert_eq!(5, dec.digit_at(0));
assert_eq!(4, dec.digit_at(1));
assert_eq!(9, dec.digit_at(2));
assert_eq!(0, dec.digit_at(3));

pub const fn magnitude_range(&self) -> RangeInclusive<i16>

Gets the visible range of digit magnitudes, in ascending order of magnitude. Call .rev() on the return value to get the range in descending order. Magnitude 0 is always included, even if the number has leading or trailing zeros.

Examples

use fixed_decimal::FixedDecimal;

let dec: FixedDecimal = "012.340".parse().expect("valid syntax");
assert_eq!(-3..=2, dec.magnitude_range());

pub fn nonzero_magnitude_left(&self) -> i16

Gets the magnitude of the largest nonzero digit. If the number is zero, 0 is returned.

Examples

use fixed_decimal::FixedDecimal;

let dec: FixedDecimal = "012.340".parse().expect("valid syntax");
assert_eq!(1, dec.nonzero_magnitude_left());

assert_eq!(0, FixedDecimal::from(0).nonzero_magnitude_left());

pub fn nonzero_magnitude_right(&self) -> i16

Gets the magnitude of the smallest nonzero digit. If the number is zero, 0 is returned.

Examples

use fixed_decimal::FixedDecimal;

let dec: FixedDecimal = "012.340".parse().expect("valid syntax");
assert_eq!(-2, dec.nonzero_magnitude_right());

assert_eq!(0, FixedDecimal::from(0).nonzero_magnitude_right());

pub fn is_zero(&self) -> bool

Returns whether the number has a numeric value of zero.

Examples

use fixed_decimal::FixedDecimal;

let dec: FixedDecimal = "000.000".parse().expect("valid syntax");
assert!(dec.is_zero());

pub fn multiply_pow10(&mut self, delta: i16) -> Result<(), Error>

Shift the digits by a power of 10, modifying self.

Leading or trailing zeros may be added to keep the digit at magnitude 0 (the last digit before the decimal separator) visible.

Can fail if the change in magnitude pushes the digits out of bounds; the magnitudes of all digits should fit in an i16.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from(42);
assert_eq!("42", dec.to_string());

dec.multiply_pow10(3).expect("Bounds are small");
assert_eq!("42000", dec.to_string());

pub fn multiplied_pow10(self, delta: i16) -> Result<Self, Error>

Shift the digits by a power of 10, consuming self and returning a new object if successful.

Leading or trailing zeros may be added to keep the digit at magnitude 0 (the last digit before the decimal separator) visible.

Can fail if the change in magnitude pushes the digits out of bounds; the magnitudes of all digits should fit in an i16.

Examples

use fixed_decimal::FixedDecimal;

let dec = FixedDecimal::from(42)
    .multiplied_pow10(3)
    .expect("Bounds are small");
assert_eq!("42000", dec.to_string());

pub fn set_sign(&mut self, sign: Sign)

Change the sign to the one given.

Examples

use fixed_decimal::FixedDecimal;
use fixed_decimal::Sign;

let mut dec = FixedDecimal::from(1729);
assert_eq!("1729", dec.to_string());

dec.set_sign(Sign::Negative);
assert_eq!("-1729", dec.to_string());

dec.set_sign(Sign::Positive);
assert_eq!("+1729", dec.to_string());

dec.set_sign(Sign::None);
assert_eq!("1729", dec.to_string());

pub fn with_sign(self, sign: Sign) -> Self

Change the sign to the one given, consuming self and returning a new object.

Examples

use fixed_decimal::FixedDecimal;
use fixed_decimal::Sign;

assert_eq!("+1729", FixedDecimal::from(1729).with_sign(Sign::Positive).to_string());
assert_eq!("1729", FixedDecimal::from(-1729).with_sign(Sign::None).to_string());
assert_eq!("-1729", FixedDecimal::from(1729).with_sign(Sign::Negative).to_string());

pub fn stripped_left(self) -> Self

Remove leading zeroes, consuming self and returning a new object.

Examples

use fixed_decimal::FixedDecimal;

let dec = FixedDecimal::from(123400)
    .multiplied_pow10(-4)
    .expect("in-bounds")
    .padded_left(4);
assert_eq!("0012.3400", dec.to_string());

assert_eq!("12.3400", dec.stripped_left().to_string());

pub fn strip_left(&mut self)

Remove leading zeroes, modifying self.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from(123400)
    .multiplied_pow10(-4)
    .expect("in-bounds")
    .padded_left(4);
assert_eq!("0012.3400", dec.to_string());

dec.strip_left();
assert_eq!("12.3400", dec.to_string());

There is no effect if the most significant digit has magnitude less than zero:

let mut dec = FixedDecimal::from(22)
    .multiplied_pow10(-4)
    .expect("in-bounds");
assert_eq!("0.0022", dec.to_string());

dec.strip_left();
assert_eq!("0.0022", dec.to_string());

pub fn stripped_right(self) -> Self

Remove trailing zeroes, consuming self and returning a new object.

Examples

use fixed_decimal::FixedDecimal;

let dec = FixedDecimal::from(123400)
    .multiplied_pow10(-4)
    .expect("in-bounds")
    .padded_left(4);
assert_eq!("0012.3400", dec.to_string());

assert_eq!("0012.34", dec.stripped_right().to_string());

pub fn strip_right(&mut self)

Remove trailing zeroes, modifying self.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from(123400)
    .multiplied_pow10(-4)
    .expect("in-bounds")
    .padded_left(4);
assert_eq!("0012.3400", dec.to_string());

dec.strip_right();
assert_eq!("0012.34", dec.to_string());

There is no effect if the least significant digit has magnitude more than zero:

let mut dec = FixedDecimal::from(2200);
assert_eq!("2200", dec.to_string());

dec.strip_right();
assert_eq!("2200", dec.to_string());

pub fn padded_left(self, position: i16) -> Self

Zero-pad the number on the left to a particular position, returning the result.

Negative numbers have no effect.

Also see FixedDecimal::truncated_left().

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from(42);
assert_eq!("42", dec.to_string());
assert_eq!("0042", dec.clone().padded_left(4).to_string());

assert_eq!("042", dec.clone().padded_left(3).to_string());

assert_eq!("42", dec.clone().padded_left(2).to_string());

assert_eq!("42", dec.clone().padded_left(1).to_string());

pub fn pad_left(&mut self, position: i16)

Zero-pad the number on the left to a particular position.

Negative numbers have no effect.

Also see FixedDecimal::truncate_left().

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from(42);
assert_eq!("42", dec.to_string());

dec.pad_left(4);
assert_eq!("0042", dec.to_string());

dec.pad_left(3);
assert_eq!("042", dec.to_string());

dec.pad_left(2);
assert_eq!("42", dec.to_string());

dec.pad_left(1);
assert_eq!("42", dec.to_string());

pub fn truncated_left(self, position: i16) -> Self

Truncate the number on the left to a particular position, deleting digits if necessary, returning the result.

Also see FixedDecimal::padded_left().

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from(4235970).multiplied_pow10(-3).expect("in-bounds");
assert_eq!("4235.970", dec.to_string());

assert_eq!("04235.970", dec.clone().truncated_left(5).to_string());

assert_eq!("35.970", dec.clone().truncated_left(2).to_string());

assert_eq!("5.970", dec.clone().truncated_left(1).to_string());

assert_eq!("0.970", dec.clone().truncated_left(0).to_string());

assert_eq!("0.070", dec.clone().truncated_left(-1).to_string());

assert_eq!("0.000", dec.clone().truncated_left(-2).to_string());

assert_eq!("0.0000", dec.clone().truncated_left(-4).to_string());

pub fn truncate_left(&mut self, position: i16)

Truncate the number on the left to a particular position, deleting digits if necessary.

Also see FixedDecimal::pad_left().

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from(4235970).multiplied_pow10(-3).expect("in-bounds");
assert_eq!("4235.970", dec.to_string());

dec.truncate_left(5);
assert_eq!("04235.970", dec.to_string());

dec.truncate_left(2);
assert_eq!("35.970", dec.to_string());

dec.truncate_left(1);
assert_eq!("5.970", dec.to_string());

dec.truncate_left(0);
assert_eq!("0.970", dec.to_string());

dec.truncate_left(-1);
assert_eq!("0.070", dec.to_string());

dec.truncate_left(-2);
assert_eq!("0.000", dec.to_string());

dec.truncate_left(-4);
assert_eq!("0.0000", dec.to_string());

pub fn truncated_right(self, position: i16) -> Self

Truncate the number on the right to a particular position, deleting digits if necessary.

Also see FixedDecimal::padded_right().

Examples

use fixed_decimal::FixedDecimal;

let dec = FixedDecimal::from_str("-1.5").unwrap();
assert_eq!("-1", dec.truncated_right(0).to_string());
let dec = FixedDecimal::from_str("0.4").unwrap();
assert_eq!("0", dec.truncated_right(0).to_string());
let dec = FixedDecimal::from_str("0.5").unwrap();
assert_eq!("0", dec.truncated_right(0).to_string());
let dec = FixedDecimal::from_str("0.6").unwrap();
assert_eq!("0", dec.truncated_right(0).to_string());
let dec = FixedDecimal::from_str("1.5").unwrap();
assert_eq!("1", dec.truncated_right(0).to_string());

pub fn truncate_right(&mut self, position: i16)

Truncates the number on the right to a particular position, deleting digits if necessary.

Also see FixedDecimal::pad_right().

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from_str("-1.5").unwrap();
dec.truncate_right(0);
assert_eq!("-1", dec.to_string());
let mut dec = FixedDecimal::from_str("0.4").unwrap();
dec.truncate_right(0);
assert_eq!("0", dec.to_string());
let mut dec = FixedDecimal::from_str("0.5").unwrap();
dec.truncate_right(0);
assert_eq!("0", dec.to_string());
let mut dec = FixedDecimal::from_str("0.6").unwrap();
dec.truncate_right(0);
assert_eq!("0", dec.to_string());
let mut dec = FixedDecimal::from_str("1.5").unwrap();
dec.truncate_right(0);
assert_eq!("1", dec.to_string());

pub fn half_truncate_right(&mut self, position: i16)

Half Truncates the number on the right to a particular position, deleting digits if necessary.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from_str("-1.5").unwrap();
dec.half_truncate_right(0);
assert_eq!("-1", dec.to_string());
let mut dec = FixedDecimal::from_str("0.4").unwrap();
dec.half_truncate_right(0);
assert_eq!("0", dec.to_string());
let mut dec = FixedDecimal::from_str("0.5").unwrap();
dec.half_truncate_right(0);
assert_eq!("0", dec.to_string());
let mut dec = FixedDecimal::from_str("0.6").unwrap();
dec.half_truncate_right(0);
assert_eq!("1", dec.to_string());   
let mut dec = FixedDecimal::from_str("1.5").unwrap();
dec.half_truncate_right(0);
assert_eq!("1", dec.to_string());
let mut dec = FixedDecimal::from_str("3.954").unwrap();
dec.half_truncate_right(0);
assert_eq!("4", dec.to_string());

pub fn half_truncated_right(self, position: i16) -> Self

Half Truncates the number on the right to a particular position, deleting digits if necessary.

Examples

use fixed_decimal::FixedDecimal;

let dec = FixedDecimal::from_str("-1.5").unwrap();
assert_eq!("-1", dec.half_truncated_right(0).to_string());
let dec = FixedDecimal::from_str("0.4").unwrap();
assert_eq!("0", dec.half_truncated_right(0).to_string());
let dec = FixedDecimal::from_str("0.5").unwrap();
assert_eq!("0", dec.half_truncated_right(0).to_string());
let dec = FixedDecimal::from_str("0.6").unwrap();
assert_eq!("1", dec.half_truncated_right(0).to_string());    
let dec = FixedDecimal::from_str("1.5").unwrap();
assert_eq!("1", dec.half_truncated_right(0).to_string());
let dec = FixedDecimal::from_str("3.954").unwrap();
assert_eq!("4", dec.half_truncated_right(0).to_string());

pub fn expand(&mut self, position: i16)

Take the expand of the number at a particular position.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from_str("-1.5").unwrap();
dec.expand(0);
assert_eq!("-2", dec.to_string());
let mut dec = FixedDecimal::from_str("0.4").unwrap();
dec.expand(0);
assert_eq!("1", dec.to_string());
let mut dec = FixedDecimal::from_str("0.5").unwrap();
dec.expand(0);
assert_eq!("1", dec.to_string());
let mut dec = FixedDecimal::from_str("0.6").unwrap();
dec.expand(0);
assert_eq!("1", dec.to_string());  
let mut dec = FixedDecimal::from_str("1.5").unwrap();
dec.expand(0);
assert_eq!("2", dec.to_string());

pub fn expanded(self, position: i16) -> Self

Take the expand of the number at a particular position.

Examples

use fixed_decimal::FixedDecimal;

let dec = FixedDecimal::from_str("-1.5").unwrap();
assert_eq!("-2", dec.expanded(0).to_string());
let dec = FixedDecimal::from_str("0.4").unwrap();
assert_eq!("1", dec.expanded(0).to_string());
let dec = FixedDecimal::from_str("0.5").unwrap();
assert_eq!("1", dec.expanded(0).to_string());
let dec = FixedDecimal::from_str("0.6").unwrap();
assert_eq!("1", dec.expanded(0).to_string());   
let dec = FixedDecimal::from_str("1.5").unwrap();
assert_eq!("2", dec.expanded(0).to_string());

pub fn half_expand(&mut self, position: i16)

Take the half expand of the number at a particular position.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from_str("-1.5").unwrap();
dec.half_expand(0);
assert_eq!("-2", dec.to_string());
let mut dec = FixedDecimal::from_str("0.4").unwrap();
dec.half_expand(0);
assert_eq!("0", dec.to_string());
let mut dec = FixedDecimal::from_str("0.5").unwrap();
dec.half_expand(0);
assert_eq!("1", dec.to_string());
let mut dec = FixedDecimal::from_str("0.6").unwrap();
dec.half_expand(0);
assert_eq!("1", dec.to_string());
let mut dec = FixedDecimal::from_str("1.5").unwrap();
dec.half_expand(0);
assert_eq!("2", dec.to_string());

pub fn half_expanded(self, position: i16) -> Self

Take the half expand of the number at a particular position.

Examples

use fixed_decimal::FixedDecimal;

let dec = FixedDecimal::from_str("-1.5").unwrap();
assert_eq!("-2", dec.half_expanded(0).to_string());
let dec = FixedDecimal::from_str("0.4").unwrap();
assert_eq!("0", dec.half_expanded(0).to_string());
let dec = FixedDecimal::from_str("0.5").unwrap();
assert_eq!("1", dec.half_expanded(0).to_string());
let dec = FixedDecimal::from_str("0.6").unwrap();
assert_eq!("1", dec.half_expanded(0).to_string());
let dec = FixedDecimal::from_str("1.5").unwrap();
assert_eq!("2", dec.half_expanded(0).to_string());

pub fn ceil(&mut self, position: i16)

Take the ceiling of the number at a particular position.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from_str("-1.5").unwrap();
dec.ceil(0);
assert_eq!("-1", dec.to_string());
let mut dec = FixedDecimal::from_str("0.4").unwrap();
dec.ceil(0);
assert_eq!("1", dec.to_string());
let mut dec = FixedDecimal::from_str("0.5").unwrap();
dec.ceil(0);
assert_eq!("1", dec.to_string());
let mut dec = FixedDecimal::from_str("0.6").unwrap();
dec.ceil(0);
assert_eq!("1", dec.to_string());
let mut dec = FixedDecimal::from_str("1.5").unwrap();
dec.ceil(0);
assert_eq!("2", dec.to_string());

pub fn ceiled(self, position: i16) -> Self

Take the ceiling of the number at a particular position.

Examples

use fixed_decimal::FixedDecimal;

let dec = FixedDecimal::from_str("-1.5").unwrap();
assert_eq!("-1", dec.ceiled(0).to_string());
let dec = FixedDecimal::from_str("0.4").unwrap();
assert_eq!("1", dec.ceiled(0).to_string());
let dec = FixedDecimal::from_str("0.5").unwrap();
assert_eq!("1", dec.ceiled(0).to_string());
let dec = FixedDecimal::from_str("0.6").unwrap();
assert_eq!("1", dec.ceiled(0).to_string());
let dec = FixedDecimal::from_str("1.5").unwrap();
assert_eq!("2", dec.ceiled(0).to_string());

pub fn half_ceil(&mut self, position: i16)

Take the half ceiling of the number at a particular position.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from_str("-1.5").unwrap();
dec.half_ceil(0);
assert_eq!("-1", dec.to_string());
let mut dec = FixedDecimal::from_str("0.4").unwrap();
dec.half_ceil(0);
assert_eq!("0", dec.to_string());
let mut dec = FixedDecimal::from_str("0.5").unwrap();
dec.half_ceil(0);
assert_eq!("1", dec.to_string());
let mut dec = FixedDecimal::from_str("0.6").unwrap();
dec.half_ceil(0);
assert_eq!("1", dec.to_string());
let mut dec = FixedDecimal::from_str("1.5").unwrap();
dec.half_ceil(0);
assert_eq!("2", dec.to_string());

pub fn half_ceiled(self, position: i16) -> Self

Take the half ceiling of the number at a particular position.

Examples

use fixed_decimal::FixedDecimal;

let dec = FixedDecimal::from_str("-1.5").unwrap();
assert_eq!("-1", dec.half_ceiled(0).to_string());
let dec = FixedDecimal::from_str("0.4").unwrap();
assert_eq!("0", dec.half_ceiled(0).to_string());
let dec = FixedDecimal::from_str("0.5").unwrap();
assert_eq!("1", dec.half_ceiled(0).to_string());
let dec = FixedDecimal::from_str("0.6").unwrap();
assert_eq!("1", dec.half_ceiled(0).to_string());
let dec = FixedDecimal::from_str("1.5").unwrap();
assert_eq!("2", dec.half_ceiled(0).to_string());

pub fn floor(&mut self, position: i16)

Take the floor of the number at a particular position.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from_str("-1.5").unwrap();
dec.floor(0);
assert_eq!("-2", dec.to_string());
let mut dec = FixedDecimal::from_str("0.4").unwrap();
dec.floor(0);
assert_eq!("0", dec.to_string());
let mut dec = FixedDecimal::from_str("0.5").unwrap();
dec.floor(0);
assert_eq!("0", dec.to_string());
let mut dec = FixedDecimal::from_str("0.6").unwrap();
dec.floor(0);
assert_eq!("0", dec.to_string());
let mut dec = FixedDecimal::from_str("1.5").unwrap();
dec.floor(0);
assert_eq!("1", dec.to_string());

pub fn floored(self, position: i16) -> Self

Take the floor of the number at a particular position.

Examples

use fixed_decimal::FixedDecimal;

let dec = FixedDecimal::from_str("-1.5").unwrap();
assert_eq!("-2", dec.floored(0).to_string());
let dec = FixedDecimal::from_str("0.4").unwrap();
assert_eq!("0", dec.floored(0).to_string());
let dec = FixedDecimal::from_str("0.5").unwrap();
assert_eq!("0", dec.floored(0).to_string());
let dec = FixedDecimal::from_str("0.6").unwrap();
assert_eq!("0", dec.floored(0).to_string());
let dec = FixedDecimal::from_str("1.5").unwrap();
assert_eq!("1", dec.floored(0).to_string());

pub fn half_floor(&mut self, position: i16)

Take the half floor of the number at a particular position.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from_str("-1.5").unwrap();
dec.half_floor(0);
assert_eq!("-2", dec.to_string());
let mut dec = FixedDecimal::from_str("0.4").unwrap();
dec.half_floor(0);
assert_eq!("0", dec.to_string());
let mut dec = FixedDecimal::from_str("0.5").unwrap();
dec.half_floor(0);
assert_eq!("0", dec.to_string());
let mut dec = FixedDecimal::from_str("0.6").unwrap();
dec.half_floor(0);
assert_eq!("1", dec.to_string());
let mut dec = FixedDecimal::from_str("1.5").unwrap();
dec.half_floor(0);
assert_eq!("1", dec.to_string());

pub fn half_floored(self, position: i16) -> Self

Take the half floor of the number at a particular position.

Examples

use fixed_decimal::FixedDecimal;

let dec = FixedDecimal::from_str("-1.5").unwrap();
assert_eq!("-2", dec.half_floored(0).to_string());
let dec = FixedDecimal::from_str("0.4").unwrap();
assert_eq!("0", dec.half_floored(0).to_string());
let dec = FixedDecimal::from_str("0.5").unwrap();
assert_eq!("0", dec.half_floored(0).to_string());
let dec = FixedDecimal::from_str("0.6").unwrap();
assert_eq!("1", dec.half_floored(0).to_string());
let dec = FixedDecimal::from_str("1.5").unwrap();
assert_eq!("1", dec.half_floored(0).to_string());

pub fn half_even(&mut self, position: i16)

Take the half even of the number at a particular position.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from_str("-1.5").unwrap();
dec.half_even(0);
assert_eq!("-2", dec.to_string());
let mut dec = FixedDecimal::from_str("0.4").unwrap();
dec.half_even(0);
assert_eq!("0", dec.to_string());
let mut dec = FixedDecimal::from_str("0.5").unwrap();
dec.half_even(0);
assert_eq!("0", dec.to_string());
let mut dec = FixedDecimal::from_str("0.6").unwrap();
dec.half_even(0);
assert_eq!("1", dec.to_string());
let mut dec = FixedDecimal::from_str("1.5").unwrap();
dec.half_even(0);
assert_eq!("2", dec.to_string());

pub fn half_evened(self, position: i16) -> Self

Take the half even of the number at a particular position.

Examples

use fixed_decimal::FixedDecimal;

let dec = FixedDecimal::from_str("-1.5").unwrap();
assert_eq!("-2", dec.half_evened(0).to_string());
let dec = FixedDecimal::from_str("0.4").unwrap();
assert_eq!("0", dec.half_evened(0).to_string());
let dec = FixedDecimal::from_str("0.5").unwrap();
assert_eq!("0", dec.half_evened(0).to_string());
let dec = FixedDecimal::from_str("0.6").unwrap();
assert_eq!("1", dec.half_evened(0).to_string());
let dec = FixedDecimal::from_str("1.5").unwrap();
assert_eq!("2", dec.half_evened(0).to_string());

pub fn padded_right(self, position: i16) -> Self

Zero-pad the number on the right to a particular (negative) position. Will truncate trailing zeros if necessary, but will not truncate other digits, returning the result.

Positive numbers have no effect.

Also see FixedDecimal::truncated_right().

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from_str("123.456").unwrap();
assert_eq!("123.456", dec.to_string());

assert_eq!("123.456", dec.clone().padded_right(-1).to_string());

assert_eq!("123.456", dec.clone().padded_right(-2).to_string());

assert_eq!("123.456000", dec.clone().padded_right(-6).to_string());

assert_eq!("123.4560", dec.clone().padded_right(-4).to_string());

pub fn pad_right(&mut self, position: i16)

Zero-pad the number on the right to a particular (negative) position. Will truncate trailing zeros if necessary, but will not truncate other digits.

Positive numbers have no effect.

Also see FixedDecimal::truncate_right().

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from_str("123.456").unwrap();
assert_eq!("123.456", dec.to_string());

dec.pad_right(-1);
assert_eq!("123.456", dec.to_string());

dec.pad_right(-2);
assert_eq!("123.456", dec.to_string());

dec.pad_right(-6);
assert_eq!("123.456000", dec.to_string());

dec.pad_right(-4);
assert_eq!("123.4560", dec.to_string());

pub fn concatenated_right(
self,
other: FixedDecimal
) -> Result<Self, FixedDecimal>

Concatenate another FixedDecimal into the end of this FixedDecimal.

All nonzero digits in other must have lower magnitude than nonzero digits in self. If the two decimals represent overlapping ranges of magnitudes, an Err is returned, passing ownership of other back to the caller.

The magnitude range of self will be increased if other covers a larger range.

Examples

use fixed_decimal::FixedDecimal;

let integer = FixedDecimal::from(123);
let fraction = FixedDecimal::from(456)
    .multiplied_pow10(-3)
    .expect("in-range");

let result =  integer.concatenated_right(fraction).expect("nonoverlapping");

assert_eq!("123.456", result.to_string());

pub fn concatenate_right(
&mut self,
other: FixedDecimal
) -> Result<(), FixedDecimal>

Concatenate another FixedDecimal into the end of this FixedDecimal.

All nonzero digits in other must have lower magnitude than nonzero digits in self. If the two decimals represent overlapping ranges of magnitudes, an Err is returned, passing ownership of other back to the caller.

The magnitude range of self will be increased if other covers a larger range.

Examples

use fixed_decimal::FixedDecimal;

let mut integer = FixedDecimal::from(123);
let fraction = FixedDecimal::from(456)
    .multiplied_pow10(-3)
    .expect("nonoverlapping");

integer.concatenate_right(fraction);

assert_eq!("123.456", integer.to_string());

pub fn signum(&self) -> Signum

Returns the Signum of this FixedDecimal.

Examples

use fixed_decimal::FixedDecimal;
use fixed_decimal::Signum;

assert_eq!(Signum::AboveZero, FixedDecimal::from(42).signum());
assert_eq!(Signum::PositiveZero, FixedDecimal::from(0).signum());
assert_eq!(
    Signum::NegativeZero,
    FixedDecimal::from_str("-0").unwrap().signum()
);
assert_eq!(Signum::BelowZero, FixedDecimal::from(-42).signum());

impl FixedDecimal

pub fn try_from_f64(
float: f64,
precision: DoublePrecision
) -> Result<Self, Error>

Construct a FixedDecimal from an f64.

Since f64 values do not carry a notion of their precision, the second argument to this function specifies the type of precision associated with the f64. For more information, see DoublePrecision.

This function uses ryu, which is an efficient double-to-string algorithm, but other implementations may yield higher performance; for more details, see icu4x#166.

This function can be made available with the "ryu" feature.

use fixed_decimal::{DoublePrecision, FixedDecimal};
use writeable::Writeable;

let decimal = FixedDecimal::try_from_f64(
    -5.1,
    DoublePrecision::Magnitude(-2),
)
.expect("Finite quantity with limited precision");
assert_eq!(decimal.write_to_string(), "-5.10");

let decimal = FixedDecimal::try_from_f64(0.012345678, DoublePrecision::Floating)
    .expect("Finite quantity");
assert_eq!(decimal.write_to_string(), "0.012345678");

let decimal = FixedDecimal::try_from_f64(12345678000., DoublePrecision::Integer)
    .expect("Finite, integer-valued quantity");
assert_eq!(decimal.write_to_string(), "12345678000");

Negative zero is supported.

use fixed_decimal::{DoublePrecision, FixedDecimal};
use writeable::Writeable;

// IEEE 754 for floating point defines the sign bit separate
// from the mantissa and exponent, allowing for -0.
let negative_zero =
    FixedDecimal::try_from_f64(-0.0, DoublePrecision::Integer).expect("Negative zero");
assert_eq!(negative_zero.write_to_string(), "-0");