fixed_decimal

Type Alias Decimal

pub type Decimal = Signed<UnsignedDecimal>;

A Type containing a UnsignedDecimal and a Sign to represent a signed decimal number.

Supports a mantissa of non-zero digits and a number of leading and trailing zeros, as well as an optional sign; used for formatting and plural selection.

Data Types

The following types can be converted to a Decimal:

Integers, signed and unsigned

Strings representing an arbitrary-precision decimal

Floating point values (using the ryu feature)

To create a Decimal with fractional digits, you have several options:

Create it from an integer and then call UnsignedDecimal::multiply_pow10 (you can also call multiply_pow10 directly on the Decimal).

Create it from a string.

When the ryu feature is enabled, create it from a floating point value using [Decimal::try_from_f64].

Examples

use fixed_decimal::Decimal;

let mut dec = Decimal::from(250);
assert_eq!("250", dec.to_string());

dec.multiply_pow10(-2);
assert_eq!("2.50", dec.to_string());

Aliased Type

struct Decimal {
    pub sign: Sign,
    pub absolute: UnsignedDecimal,
}

Fields
sign: Signabsolute: UnsignedDecimal
Implementations

impl Decimal

pub fn new(sign: Sign, absolute: UnsignedDecimal) -> Self

pub fn try_from_str(s: &str) -> Result<Self, ParseError>

Parses a Decimal.

pub fn try_from_utf8(input_str: &[u8]) -> Result<Self, ParseError>

pub fn apply_sign_display(&mut self, sign_display: SignDisplay)

Sets the sign of this number according to the given sign display strategy.

Examples

use fixed_decimal::Decimal;
use fixed_decimal::SignDisplay::*;

let mut dec = Decimal::from(1729);
assert_eq!("1729", dec.to_string());
dec.apply_sign_display(Always);
assert_eq!("+1729", dec.to_string());

pub fn with_sign_display(self, sign_display: SignDisplay) -> Self

Returns this number with its sign set according to the given sign display strategy.

Examples

use fixed_decimal::Decimal;
use fixed_decimal::SignDisplay::*;

assert_eq!(
    "+1729",
    Decimal::from(1729)
        .with_sign_display(ExceptZero)
        .to_string()
);

impl Decimal

All the rounding and rounding related logic is implmented in this implmentation block.

pub fn round(&mut self, position: i16)

Rounds this number at a particular digit position.

This uses half to even rounding, which rounds to the nearest integer and resolves ties by selecting the nearest even integer to the original value.

Examples

use fixed_decimal::Decimal;

let mut dec = Decimal::from_str("-1.5").unwrap();
dec.round(0);
assert_eq!("-2", dec.to_string());
let mut dec = Decimal::from_str("0.4").unwrap();
dec.round(0);
assert_eq!("0", dec.to_string());
let mut dec = Decimal::from_str("0.5").unwrap();
dec.round(0);
assert_eq!("0", dec.to_string());
let mut dec = Decimal::from_str("0.6").unwrap();
dec.round(0);
assert_eq!("1", dec.to_string());
let mut dec = Decimal::from_str("1.5").unwrap();
dec.round(0);
assert_eq!("2", dec.to_string());

pub fn rounded(self, position: i16) -> Self

Returns this number rounded at a particular digit position.

This uses half to even rounding by default, which rounds to the nearest integer and resolves ties by selecting the nearest even integer to the original value.

Examples

use fixed_decimal::Decimal;

let mut dec = Decimal::from_str("-1.5").unwrap();
assert_eq!("-2", dec.rounded(0).to_string());
let mut dec = Decimal::from_str("0.4").unwrap();
assert_eq!("0", dec.rounded(0).to_string());
let mut dec = Decimal::from_str("0.5").unwrap();
assert_eq!("0", dec.rounded(0).to_string());
let mut dec = Decimal::from_str("0.6").unwrap();
assert_eq!("1", dec.rounded(0).to_string());
let mut dec = Decimal::from_str("1.5").unwrap();
assert_eq!("2", dec.rounded(0).to_string());

pub fn ceil(&mut self, position: i16)

Rounds this number towards positive infinity at a particular digit position.

Examples

use fixed_decimal::Decimal;

let mut dec = Decimal::from_str("-1.5").unwrap();
dec.ceil(0);
assert_eq!("-1", dec.to_string());
let mut dec = Decimal::from_str("0.4").unwrap();
dec.ceil(0);
assert_eq!("1", dec.to_string());
let mut dec = Decimal::from_str("0.5").unwrap();
dec.ceil(0);
assert_eq!("1", dec.to_string());
let mut dec = Decimal::from_str("0.6").unwrap();
dec.ceil(0);
assert_eq!("1", dec.to_string());
let mut dec = Decimal::from_str("1.5").unwrap();
dec.ceil(0);
assert_eq!("2", dec.to_string());

pub fn ceiled(self, position: i16) -> Self

Returns this number rounded towards positive infinity at a particular digit position.

Examples

use fixed_decimal::Decimal;

let dec = Decimal::from_str("-1.5").unwrap();
assert_eq!("-1", dec.ceiled(0).to_string());
let dec = Decimal::from_str("0.4").unwrap();
assert_eq!("1", dec.ceiled(0).to_string());
let dec = Decimal::from_str("0.5").unwrap();
assert_eq!("1", dec.ceiled(0).to_string());
let dec = Decimal::from_str("0.6").unwrap();
assert_eq!("1", dec.ceiled(0).to_string());
let dec = Decimal::from_str("1.5").unwrap();
assert_eq!("2", dec.ceiled(0).to_string());

pub fn expand(&mut self, position: i16)

Rounds this number away from zero at a particular digit position.

Examples

use fixed_decimal::Decimal;

let mut dec = Decimal::from_str("-1.5").unwrap();
dec.expand(0);
assert_eq!("-2", dec.to_string());
let mut dec = Decimal::from_str("0.4").unwrap();
dec.expand(0);
assert_eq!("1", dec.to_string());
let mut dec = Decimal::from_str("0.5").unwrap();
dec.expand(0);
assert_eq!("1", dec.to_string());
let mut dec = Decimal::from_str("0.6").unwrap();
dec.expand(0);
assert_eq!("1", dec.to_string());
let mut dec = Decimal::from_str("1.5").unwrap();
dec.expand(0);
assert_eq!("2", dec.to_string());

pub fn expanded(self, position: i16) -> Self

Returns this number rounded away from zero at a particular digit position.

Examples

use fixed_decimal::Decimal;

let dec = Decimal::from_str("-1.5").unwrap();
assert_eq!("-2", dec.expanded(0).to_string());
let dec = Decimal::from_str("0.4").unwrap();
assert_eq!("1", dec.expanded(0).to_string());
let dec = Decimal::from_str("0.5").unwrap();
assert_eq!("1", dec.expanded(0).to_string());
let dec = Decimal::from_str("0.6").unwrap();
assert_eq!("1", dec.expanded(0).to_string());
let dec = Decimal::from_str("1.5").unwrap();
assert_eq!("2", dec.expanded(0).to_string());

pub fn floor(&mut self, position: i16)

Rounds this number towards negative infinity at a particular digit position.

Examples

use fixed_decimal::Decimal;

let mut dec = Decimal::from_str("-1.5").unwrap();
dec.floor(0);
assert_eq!("-2", dec.to_string());
let mut dec = Decimal::from_str("0.4").unwrap();
dec.floor(0);
assert_eq!("0", dec.to_string());
let mut dec = Decimal::from_str("0.5").unwrap();
dec.floor(0);
assert_eq!("0", dec.to_string());
let mut dec = Decimal::from_str("0.6").unwrap();
dec.floor(0);
assert_eq!("0", dec.to_string());
let mut dec = Decimal::from_str("1.5").unwrap();
dec.floor(0);
assert_eq!("1", dec.to_string());

pub fn floored(self, position: i16) -> Self

Returns this number rounded towards negative infinity at a particular digit position.

Examples

use fixed_decimal::Decimal;

let dec = Decimal::from_str("-1.5").unwrap();
assert_eq!("-2", dec.floored(0).to_string());
let dec = Decimal::from_str("0.4").unwrap();
assert_eq!("0", dec.floored(0).to_string());
let dec = Decimal::from_str("0.5").unwrap();
assert_eq!("0", dec.floored(0).to_string());
let dec = Decimal::from_str("0.6").unwrap();
assert_eq!("0", dec.floored(0).to_string());
let dec = Decimal::from_str("1.5").unwrap();
assert_eq!("1", dec.floored(0).to_string());

pub fn trunc(&mut self, position: i16)

Rounds this number towards zero at a particular digit position.

Also see UnsignedDecimal::pad_end().

Examples

use fixed_decimal::Decimal;

let mut dec = Decimal::from_str("-1.5").unwrap();
dec.trunc(0);
assert_eq!("-1", dec.to_string());
let mut dec = Decimal::from_str("0.4").unwrap();
dec.trunc(0);
assert_eq!("0", dec.to_string());
let mut dec = Decimal::from_str("0.5").unwrap();
dec.trunc(0);
assert_eq!("0", dec.to_string());
let mut dec = Decimal::from_str("0.6").unwrap();
dec.trunc(0);
assert_eq!("0", dec.to_string());
let mut dec = Decimal::from_str("1.5").unwrap();
dec.trunc(0);
assert_eq!("1", dec.to_string());

pub fn trunced(self, position: i16) -> Self

Returns this number rounded towards zero at a particular digit position.

Also see UnsignedDecimal::padded_end().

Examples

use fixed_decimal::Decimal;

let dec = Decimal::from_str("-1.5").unwrap();
assert_eq!("-1", dec.trunced(0).to_string());
let dec = Decimal::from_str("0.4").unwrap();
assert_eq!("0", dec.trunced(0).to_string());
let dec = Decimal::from_str("0.5").unwrap();
assert_eq!("0", dec.trunced(0).to_string());
let dec = Decimal::from_str("0.6").unwrap();
assert_eq!("0", dec.trunced(0).to_string());
let dec = Decimal::from_str("1.5").unwrap();
assert_eq!("1", dec.trunced(0).to_string());

pub fn round_with_mode(&mut self, position: i16, mode: SignedRoundingMode)

Rounds this number at a particular digit position, using the specified rounding mode.

Examples

use fixed_decimal::{Decimal, SignedRoundingMode, UnsignedRoundingMode};

let mut dec = Decimal::from_str("-3.5").unwrap();
dec.round_with_mode(0, SignedRoundingMode::Floor);
assert_eq!("-4", dec.to_string());
let mut dec = Decimal::from_str("-3.5").unwrap();
dec.round_with_mode(0, SignedRoundingMode::Ceil);
assert_eq!("-3", dec.to_string());
let mut dec = Decimal::from_str("5.455").unwrap();
dec.round_with_mode(
    -2,
    SignedRoundingMode::Unsigned(UnsignedRoundingMode::HalfExpand),
);
assert_eq!("5.46", dec.to_string());
let mut dec = Decimal::from_str("-7.235").unwrap();
dec.round_with_mode(
    -2,
    SignedRoundingMode::Unsigned(UnsignedRoundingMode::HalfTrunc),
);
assert_eq!("-7.23", dec.to_string());
let mut dec = Decimal::from_str("9.75").unwrap();
dec.round_with_mode(
    -1,
    SignedRoundingMode::Unsigned(UnsignedRoundingMode::HalfEven),
);
assert_eq!("9.8", dec.to_string());

pub fn rounded_with_mode(self, position: i16, mode: SignedRoundingMode) -> Self

Returns this number rounded at a particular digit position, using the specified rounding mode.

Examples

use fixed_decimal::{Decimal, SignedRoundingMode, UnsignedRoundingMode};

let mut dec = Decimal::from_str("-3.5").unwrap();
assert_eq!(
    "-4",
    dec.rounded_with_mode(0, SignedRoundingMode::Floor)
        .to_string()
);
let mut dec = Decimal::from_str("-3.5").unwrap();
assert_eq!(
    "-3",
    dec.rounded_with_mode(0, SignedRoundingMode::Ceil)
        .to_string()
);
let mut dec = Decimal::from_str("5.455").unwrap();
assert_eq!(
    "5.46",
    dec.rounded_with_mode(
        -2,
        SignedRoundingMode::Unsigned(UnsignedRoundingMode::HalfExpand)
    )
    .to_string()
);
let mut dec = Decimal::from_str("-7.235").unwrap();
assert_eq!(
    "-7.23",
    dec.rounded_with_mode(
        -2,
        SignedRoundingMode::Unsigned(UnsignedRoundingMode::HalfTrunc)
    )
    .to_string()
);
let mut dec = Decimal::from_str("9.75").unwrap();
assert_eq!(
    "9.8",
    dec.rounded_with_mode(
        -1,
        SignedRoundingMode::Unsigned(UnsignedRoundingMode::HalfEven)
    )
    .to_string()
);

pub fn round_with_mode_and_increment( &mut self, position: i16, mode: SignedRoundingMode, increment: RoundingIncrement, )

Rounds this number at a particular digit position and increment, using the specified rounding mode.

Examples

use fixed_decimal::{
    Decimal, RoundingIncrement, SignedRoundingMode, UnsignedRoundingMode,
};

let mut dec = Decimal::from_str("-3.5").unwrap();
dec.round_with_mode_and_increment(
    0,
    SignedRoundingMode::Floor,
    RoundingIncrement::MultiplesOf1,
);
assert_eq!("-4", dec.to_string());
let mut dec = Decimal::from_str("-3.59").unwrap();
dec.round_with_mode_and_increment(
    -1,
    SignedRoundingMode::Ceil,
    RoundingIncrement::MultiplesOf2,
);
assert_eq!("-3.4", dec.to_string());
let mut dec = Decimal::from_str("5.455").unwrap();
dec.round_with_mode_and_increment(
    -2,
    SignedRoundingMode::Unsigned(UnsignedRoundingMode::HalfExpand),
    RoundingIncrement::MultiplesOf5,
);
assert_eq!("5.45", dec.to_string());
let mut dec = Decimal::from_str("-7.235").unwrap();
dec.round_with_mode_and_increment(
    -2,
    SignedRoundingMode::Unsigned(UnsignedRoundingMode::HalfTrunc),
    RoundingIncrement::MultiplesOf25,
);
assert_eq!("-7.25", dec.to_string());
let mut dec = Decimal::from_str("9.75").unwrap();
dec.round_with_mode_and_increment(
    -1,
    SignedRoundingMode::Unsigned(UnsignedRoundingMode::HalfEven),
    RoundingIncrement::MultiplesOf5,
);
assert_eq!("10.0", dec.to_string());

pub fn rounded_with_mode_and_increment( self, position: i16, mode: SignedRoundingMode, increment: RoundingIncrement, ) -> Self

Returns this number rounded at a particular digit position and increment, using the specified rounding mode.

Examples

use fixed_decimal::{
    Decimal, RoundingIncrement, SignedRoundingMode, UnsignedRoundingMode,
};

let mut dec = Decimal::from_str("-3.5").unwrap();
assert_eq!(
    "-4",
    dec.rounded_with_mode_and_increment(
        0,
        SignedRoundingMode::Floor,
        RoundingIncrement::MultiplesOf1
    )
    .to_string()
);
let mut dec = Decimal::from_str("-3.59").unwrap();
assert_eq!(
    "-3.4",
    dec.rounded_with_mode_and_increment(
        -1,
        SignedRoundingMode::Ceil,
        RoundingIncrement::MultiplesOf2
    )
    .to_string()
);
let mut dec = Decimal::from_str("5.455").unwrap();
assert_eq!(
    "5.45",
    dec.rounded_with_mode_and_increment(
        -2,
        SignedRoundingMode::Unsigned(UnsignedRoundingMode::HalfExpand),
        RoundingIncrement::MultiplesOf5
    )
    .to_string()
);
let mut dec = Decimal::from_str("-7.235").unwrap();
assert_eq!(
    "-7.25",
    dec.rounded_with_mode_and_increment(
        -2,
        SignedRoundingMode::Unsigned(UnsignedRoundingMode::HalfTrunc),
        RoundingIncrement::MultiplesOf25
    )
    .to_string()
);
let mut dec = Decimal::from_str("9.75").unwrap();
assert_eq!(
    "10.0",
    dec.rounded_with_mode_and_increment(
        -1,
        SignedRoundingMode::Unsigned(UnsignedRoundingMode::HalfEven),
        RoundingIncrement::MultiplesOf5
    )
    .to_string()
);

impl Decimal

pub fn to_string(&self) -> String

Converts the given value to a String.

Under the hood, this uses an efficient [Writeable] implementation. However, in order to avoid allocating a string, it is more efficient to use [Writeable] directly.