Struct fixed_decimal::decimal::FixedDecimal

pub struct FixedDecimal { /* private fields */ }

A struct containing decimal digits with efficient iteration and manipulation by magnitude (power of 10). Supports a mantissa of non-zero digits and a number of leading and trailing zeros, used for formatting and plural selection.

Data Types

The following types can be converted to a FixedDecimal:

Integers, signed and unsigned

Strings representing an arbitrary-precision decimal

To create a FixedDecimal with fraction digits, either create it from an integer and then call FixedDecimal::multiplied_pow10, or create it from a string.

Floating point numbers will be supported pending a resolution to #166. In the mean time, a third-party float-to-string library may be used.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from(250);
assert_eq!("250", dec.to_string());

dec.multiply_pow10(-2);
assert_eq!("2.50", dec.to_string());

Implementations

impl FixedDecimal

pub fn digit_at(&self, magnitude: i16) -> u8

Gets the digit at the specified order of magnitude. Returns 0 if the magnitude is out of range of the currently visible digits.

Examples

use fixed_decimal::FixedDecimal;

let dec = FixedDecimal::from(945);
assert_eq!(0, dec.digit_at(-1));
assert_eq!(5, dec.digit_at(0));
assert_eq!(4, dec.digit_at(1));
assert_eq!(9, dec.digit_at(2));
assert_eq!(0, dec.digit_at(3));

pub const fn magnitude_range(&self) -> RangeInclusive<i16>

Gets the visible range of digit magnitudes, in ascending order of magnitude. Call .rev() on the return value to get the range in descending order. Magnitude 0 is always included, even if the number has leading or trailing zeros.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from(120);
assert_eq!(0..=2, dec.magnitude_range());

pub fn multiply_pow10(&mut self, delta: i16) -> Result<(), Error>

Shift the digits by a power of 10, modifying self.

Leading or trailing zeros may be added to keep the digit at magnitude 0 (the last digit before the decimal separator) visible.

Can fail if the change in magnitude pushes the digits out of bounds; the magnitudes of all digits should fit in an i16.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from(42);
assert_eq!("42", dec.to_string());

dec.multiply_pow10(3).expect("Bounds are small");
assert_eq!("42000", dec.to_string());

pub fn multiplied_pow10(self, delta: i16) -> Result<Self, Error>

Shift the digits by a power of 10, consuming self and returning a new object if successful.

Leading or trailing zeros may be added to keep the digit at magnitude 0 (the last digit before the decimal separator) visible.

Can fail if the change in magnitude pushes the digits out of bounds; the magnitudes of all digits should fit in an i16.

Examples

use fixed_decimal::FixedDecimal;

let dec = FixedDecimal::from(42).multiplied_pow10(3).expect("Bounds are small");
assert_eq!("42000", dec.to_string());

pub fn negate(&mut self)

Change the value from negative to positive or from positive to negative, modifying self. Negative zero is supported.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from(42);
assert_eq!("42", dec.to_string());

dec.negate();
assert_eq!("-42", dec.to_string());

dec.negate();
assert_eq!("42", dec.to_string());

// Negative zero example
let zero = FixedDecimal::from(0);
let mut negative_zero = FixedDecimal::from(0);
negative_zero.negate();

assert_eq!("0", zero.to_string());
assert_eq!("-0", negative_zero.to_string());
assert_ne!(zero, negative_zero);

pub fn negated(self) -> Self

Change the value from negative to positive or from positive to negative, consuming self and returning a new object.

Examples

use fixed_decimal::FixedDecimal;

assert_eq!(FixedDecimal::from(-42), FixedDecimal::from(42).negated());

pub fn padded_left(self, digits: u16) -> Self

Zero-pad the number on the left to a particular number of integer digits, returning the result.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from(42);
assert_eq!("42", dec.to_string());

;
assert_eq!("0042", dec.clone().padded_left(4).to_string());

assert_eq!("042", dec.clone().padded_left(3).to_string());

assert_eq!("42", dec.clone().padded_left(2).to_string());

assert_eq!("42", dec.clone().padded_left(1).to_string());

pub fn pad_left(&mut self, digits: u16)

Zero-pad the number on the left to a particular number of integer digits.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from(42);
assert_eq!("42", dec.to_string());

dec.pad_left(4);
assert_eq!("0042", dec.to_string());

dec.pad_left(3);
assert_eq!("042", dec.to_string());

dec.pad_left(2);
assert_eq!("42", dec.to_string());

dec.pad_left(1);
assert_eq!("42", dec.to_string());

pub fn truncated_left(self, magnitude: i16) -> Self

Truncate the number on the left to a particular magnitude, deleting digits if necessary, returning the result.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from(4235);
assert_eq!("4235", dec.to_string());

assert_eq!("4235", dec.clone().truncated_left(5).to_string());

assert_eq!("235", dec.clone().truncated_left(2).to_string());

assert_eq!("35", dec.clone().truncated_left(1).to_string());

assert_eq!("5", dec.clone().truncated_left(0).to_string());

assert_eq!("0", dec.clone().truncated_left(-1).to_string());

pub fn truncate_left(&mut self, magnitude: i16)

Truncate the number on the left to a particular magnitude, deleting digits if necessary.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from(4235);
assert_eq!("4235", dec.to_string());

dec.truncate_left(5);
assert_eq!("4235", dec.to_string());

dec.truncate_left(2);
assert_eq!("235", dec.to_string());

dec.truncate_left(1);
assert_eq!("35", dec.to_string());

dec.truncate_left(0);
assert_eq!("5", dec.to_string());

dec.truncate_left(-1);
assert_eq!("0", dec.to_string());

pub fn padded_right(self, negative_magnitude: u16) -> Self

Zero-pad the number on the right to a particular (negative) magnitude. Will truncate trailing zeros if necessary, but will not truncate other digits, returning the result.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from_str("123.456").unwrap();
assert_eq!("123.456", dec.to_string());

assert_eq!("123.456", dec.clone().padded_right(1).to_string());

assert_eq!("123.456", dec.clone().padded_right(2).to_string());

assert_eq!("123.4560", dec.clone().padded_right(4).to_string());

assert_eq!("123.456000", dec.clone().padded_right(6).to_string());

pub fn pad_right(&mut self, negative_magnitude: u16)

Zero-pad the number on the right to a particular (negative) magnitude. Will truncate trailing zeros if necessary, but will not truncate other digits.

Examples

use fixed_decimal::FixedDecimal;

let mut dec = FixedDecimal::from_str("123.456").unwrap();
assert_eq!("123.456", dec.to_string());

dec.pad_right(1);
assert_eq!("123.456", dec.to_string());

dec.pad_right(2);
assert_eq!("123.456", dec.to_string());

dec.pad_right(4);
assert_eq!("123.4560", dec.to_string());

dec.pad_right(6);
assert_eq!("123.456000", dec.to_string());

pub fn signum(&self) -> Signum

Returns the Signum of this FixedDecimal.

Examples

use fixed_decimal::FixedDecimal;
use fixed_decimal::Signum;

assert_eq!(Signum::AboveZero, FixedDecimal::from(42).signum());
assert_eq!(Signum::PositiveZero, FixedDecimal::from(0).signum());
assert_eq!(Signum::NegativeZero, FixedDecimal::from(0).negated().signum());
assert_eq!(Signum::BelowZero, FixedDecimal::from(-42).signum());